- add request validation

+++ ImaginaryTaxCalculator/ImaginaryTaxCalculator.Service/Models/TaxPayer.cs
@@ -3,7 +3,7 @@
     public class TaxPayer
     {
         public string FullName { get; set; }
-        public int SSN { get; set; }
+        public long SSN { get; set; }
         public decimal GrossIncome { get; set; }
         public decimal? CharitySpent { get; set; }
